pyproject.toml
[project]
name = "daily-photo-bot"
version = "0.2.3"
description = "Telegram bot that sends daily photos based on user-chosen topics"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot[job-queue]>=21.0",
    "aiosqlite>=0.20.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "ruff>=0.5.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
